refactor: 파일,폴더 정리 및 컨벤션통일
