+++ Diary/Model/Support/CacheManager.swift
@@ -8,6 +8,8 @@
 import UIKit
 
 final class ImageCacheManager {
+    
     static let shared = NSCache<NSString, UIImage>()
+    
     private init() {}
 }
